publish_profile_create_function/requirements.txt
functions_framework
google-cloud-bigquery
google-cloud-pubsub
pytz
python-dotenv

